Feature: Updates planner install script so it can run multiple times.

+++ skiros2/scripts/install_fd_task_planner.sh
@@ -5,7 +5,7 @@ tfd='tfd-src-0.4'
 echo "Installing planner..."
 
 #Navigate to install folder
-default="~/utils"
+default="~/Software"
 echo "Select install folder or leave blank for default [default: $default]:"
 read folder
 if [[ $folder == "" ]]; then
@@ -13,30 +13,38 @@ if [[ $folder == "" ]]; then
 else
     folder=${folder/"~"/$HOME}
 fi
-mkdir -p $folder
-cd $folder
-echo $folder
 
-#Install
-wget "http://gki.informatik.uni-freiburg.de/tools/tfd/downloads/version-0.4/${tfd}.tgz"
-tar xzf "${tfd}.tgz"
-cd "${tfd}"
-sed -e s/"-Werror"//g -i ./downward/search/Makefile
-sed -e s/"translate\/"//g -i ./downward/plan.py
-sed -e s/"preprocess\/"//g -i ./downward/plan.py
-sed -e s/"search\/"//g -i ./downward/plan.py
-./build
-cd -
-rm -r "${tfd}.tgz"
+# Installing planner
+if [ ! -d "$folder/${tfd}" ]; then
+    echo "Installing planner."
+    mkdir -p $folder
+    cd $folder
+    echo $folder
+    wget "http://gki.informatik.uni-freiburg.de/tools/tfd/downloads/version-0.4/${tfd}.tgz"
+    tar xzf "${tfd}.tgz"
+    cd "${tfd}"
+    sed -e s/"-Werror"//g -i ./downward/search/Makefile
+    sed -e s/"translate\/"//g -i ./downward/plan.py
+    sed -e s/"preprocess\/"//g -i ./downward/plan.py
+    sed -e s/"search\/"//g -i ./downward/plan.py
+    ./build
+    cd -
+    rm -r "${tfd}.tgz"
+else
+	echo "Folder $folder/${tfd} already exists. Skipping installation."
+fi
 
-#Add environment variable to bashrc
-string="export TFD_HOME=$(pwd)"
+# Adding environment variables
+echo "Add environment variable to bashrc"
+string="export TFD_HOME=${folder}"
 string2="export PATH=$""TFD_HOME/${tfd}/downward:$""TFD_HOME/${tfd}/downward/translate:$""TFD_HOME/${tfd}/downward/preprocess:$""TFD_HOME/${tfd}/downward/search:""$""PATH"
 temp=$(cat ~/.bashrc | grep "$string")
 if [ -z "$temp" ]; then
+    echo "# TDF Planner for SkiROS" >> ~/.bashrc
 	echo $string >> ~/.bashrc
 	echo $string2 >> ~/.bashrc
 	echo "Printing string in .bashrc: $string"
+	echo "Printing string in .bashrc: $string2"
 else
 	echo "No export string added to bashrc (it is already there)"
 fi
